+++ src/main/java/org/thisdote/communityfordevelopers/user/dto/UserDTO.java
@@ -1,0 +1,24 @@
+package org.thisdote.communityfordevelopers.user.dto;
+
+import lombok.*;
+
+import java.util.Date;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@ToString
+public class UserDTO {
+    private int userCode;
+    private String userId;
+    private String userPassword;
+    private Date userBirthday;
+    private String userPhone;
+    private String userEmail;
+    private int userStudyGroupStatus;
+    private Date userRegistDate;
+    private Date userInfoUpdateDate;
+    private int userGrade;
+//    private StudyGroupMemberDTO studyGroupMemberDTO;
+}
