+++ src/main/java/org/thisdote/communityfordevelopers/user/query/UserMapper.java
@@ -1,6 +1,7 @@
 package org.thisdote.communityfordevelopers.user.query;
 
 import org.apache.ibatis.annotations.Mapper;
+import org.thisdote.communityfordevelopers.user.dto.UserDTO;
 
 import java.util.List;
